requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
sniffio==1.3.0
SQLAlchemy==2.0.31
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
